Fixed type conversion warning

+++ RogaliqueGame/Ball.cpp
@@ -13,7 +13,7 @@ namespace
 namespace RogaliqueGame
 {
 	Ball::Ball(const sf::Vector2f& position)
-		: GameObject(SETTINGS.TEXTURES_PATH + TEXTURE_ID + ".png", position, SETTINGS.BALL_SIZE, SETTINGS.BALL_SIZE)
+		: GameObject(SETTINGS.TEXTURES_PATH + TEXTURE_ID + ".png", position, (float)SETTINGS.BALL_SIZE, (float)SETTINGS.BALL_SIZE)
 	{
 		const float angle = 90;
 		const auto pi = std::acos(-1.f);
