+++ RogaliqueGame/LevelLoader.cpp
@@ -58,6 +58,6 @@ namespace RogaliqueGame {
 
 	int LevelLoader::GetLevelCount()
 	{
-		return levels.size();
+		return (int)levels.size();
 	}
 }
